split AppDelegate in extensions
